+++ android/src/main/java/com/setupadprebidreactnative/VeonPrebidReactNativeViewManager.kt
@@ -1,17 +1,16 @@
 package com.setupadprebidreactnative
 
-import android.graphics.Color
 import android.util.Log
-import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReadableArray
-import com.facebook.react.bridge.WritableMap
-import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.bridge.UiThreadUtil
+import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.uimanager.SimpleViewManager
 import com.facebook.react.uimanager.ThemedReactContext
+import com.facebook.react.uimanager.ViewManagerDelegate
 import com.facebook.react.uimanager.annotations.ReactProp
-import com.facebook.react.uimanager.events.RCTEventEmitter
+import com.facebook.react.viewmanagers.VeonPrebidReactNativeViewManagerDelegate
+import com.facebook.react.viewmanagers.VeonPrebidReactNativeViewManagerInterface
 
 /**
  * ViewManager for Prebid ad views
@@ -20,9 +19,14 @@ import com.facebook.react.uimanager.events.RCTEventEmitter
 @ReactModule(name = VeonPrebidReactNativeViewManager.NAME)
 class VeonPrebidReactNativeViewManager(
     private val reactContext: ReactApplicationContext
-) : SimpleViewManager<VeonPrebidReactNativeView>() {
+) : SimpleViewManager<VeonPrebidReactNativeView>(),
+    VeonPrebidReactNativeViewManagerInterface<VeonPrebidReactNativeView> {
 
     private val TAG = "VeonPrebidViewManager"
+
+    private val delegate = VeonPrebidReactNativeViewManagerDelegate(this)
+
+    override fun getDelegate(): ViewManagerDelegate<VeonPrebidReactNativeView> = delegate
 
     override fun getName(): String = NAME
 
@@ -31,38 +35,40 @@ class VeonPrebidReactNativeViewManager(
         return VeonPrebidReactNativeView(context)
     }
 
+    // region Props
+
     /**
      * Set ad type (banner, interstitial, rewardVideo)
      */
     @ReactProp(name = "adType")
-    fun setAdType(view: VeonPrebidReactNativeView, adType: String) {
+    override fun setAdType(view: VeonPrebidReactNativeView, adType: String?) {
         Log.d(TAG, "Setting ad type: $adType")
-        view.setAdType(adType)
+        view.setAdType(adType ?: return)
     }
 
     /**
      * Set Prebid config ID
      */
     @ReactProp(name = "configId")
-    fun setConfigId(view: VeonPrebidReactNativeView, configId: String) {
+    override fun setConfigId(view: VeonPrebidReactNativeView, configId: String?) {
         Log.d(TAG, "Setting config ID: $configId")
-        view.setConfigId(configId)
+        view.setConfigId(configId ?: return)
     }
 
     /**
      * Set Google Ad Manager ad unit ID
      */
     @ReactProp(name = "adUnitId")
-    fun setAdUnitId(view: VeonPrebidReactNativeView, adUnitId: String) {
+    override fun setAdUnitId(view: VeonPrebidReactNativeView, adUnitId: String?) {
         Log.d(TAG, "Setting ad unit ID: $adUnitId")
-        view.setAdUnitId(adUnitId)
+        view.setAdUnitId(adUnitId ?: return)
     }
 
     /**
      * Set ad width
      */
     @ReactProp(name = "width")
-    fun setWidth(view: VeonPrebidReactNativeView, width: Int) {
+    override fun setWidth(view: VeonPrebidReactNativeView, width: Int) {
         Log.d(TAG, "Setting width: $width")
         view.setAdWidth(width)
     }
@@ -71,7 +77,7 @@ class VeonPrebidReactNativeViewManager(
      * Set ad height
      */
     @ReactProp(name = "height")
-    fun setHeight(view: VeonPrebidReactNativeView, height: Int) {
+    override fun setHeight(view: VeonPrebidReactNativeView, height: Int) {
         Log.d(TAG, "Setting height: $height")
         view.setAdHeight(height)
     }
@@ -80,62 +86,57 @@ class VeonPrebidReactNativeViewManager(
      * Set refresh interval in seconds (for banners)
      */
     @ReactProp(name = "refreshInterval")
-    fun setRefreshInterval(view: VeonPrebidReactNativeView, refreshInterval: Int) {
+    override fun setRefreshInterval(view: VeonPrebidReactNativeView, refreshInterval: Int) {
         Log.d(TAG, "Setting refresh interval: $refreshInterval seconds")
         view.setRefreshInterval(refreshInterval)
     }
 
-    /**
-     * Define commands that can be called from JavaScript
-     */
-    override fun getCommandsMap(): Map<String, Int> = mapOf(
-        "loadBanner" to COMMAND_LOAD_BANNER,
-        "showBanner" to COMMAND_SHOW_BANNER,
-        "hideBanner" to COMMAND_HIDE_BANNER,
-        "loadInterstitial" to COMMAND_LOAD_INTERSTITIAL,
-        "showInterstitial" to COMMAND_SHOW_INTERSTITIAL,
-        "hideInterstitial" to COMMAND_HIDE_INTERSTITIAL,
-        "pauseAuction" to COMMAND_PAUSE_AUCTION,
-        "resumeAuction" to COMMAND_RESUME_AUCTION,
-        "destroyAuction" to COMMAND_DESTROY_AUCTION
-    )
+    override fun loadBanner(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.loadBanner() }
+    }
 
-    /**
-     * Handle commands from JavaScript (new architecture - String command)
-     */
-    override fun receiveCommand(
-        view: VeonPrebidReactNativeView,
-        commandId: String,
-        args: ReadableArray?
-    ) {
-        Log.d(TAG, "Received command (String): $commandId")
+    override fun showBanner(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.showBanner() }
+    }
 
-        UiThreadUtil.runOnUiThread {
-            when (commandId) {
-                "loadBanner", COMMAND_LOAD_BANNER.toString() -> view.loadBanner()
-                "showBanner", COMMAND_SHOW_BANNER.toString() -> view.showBanner()
-                "hideBanner", COMMAND_HIDE_BANNER.toString() -> view.hideBanner()
-                "loadInterstitial", COMMAND_LOAD_INTERSTITIAL.toString() -> view.loadInterstitial()
-                "showInterstitial", COMMAND_SHOW_INTERSTITIAL.toString() -> view.showInterstitial()
-                "hideInterstitial", COMMAND_HIDE_INTERSTITIAL.toString() -> view.hideInterstitial()
-                "pauseAuction", COMMAND_PAUSE_AUCTION.toString() -> view.pauseAuction()
-                "resumeAuction", COMMAND_RESUME_AUCTION.toString() -> view.resumeAuction()
-                "destroyAuction", COMMAND_DESTROY_AUCTION.toString() -> view.destroyAuction()
-                else -> Log.w(TAG, "Unknown command: $commandId")
-            }
-        }
+    override fun hideBanner(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.hideBanner() }
+    }
+
+    override fun loadInterstitial(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.loadInterstitial() }
+    }
+
+    override fun showInterstitial(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.showInterstitial() }
+    }
+
+    override fun hideInterstitial(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.hideInterstitial() }
+    }
+
+    override fun pauseAuction(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.pauseAuction() }
+    }
+
+    override fun resumeAuction(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.resumeAuction() }
+    }
+
+    override fun destroyAuction(view: VeonPrebidReactNativeView) {
+        UiThreadUtil.runOnUiThread { view.destroyAuction() }
     }
 
     /**
      * Handle commands from JavaScript (old architecture - Int command)
      */
+    @Deprecated("Deprecated in Java")
     override fun receiveCommand(
         view: VeonPrebidReactNativeView,
         commandId: Int,
         args: ReadableArray?
     ) {
         Log.d(TAG, "Received command (Int): $commandId")
-
         UiThreadUtil.runOnUiThread {
             when (commandId) {
                 COMMAND_LOAD_BANNER -> view.loadBanner()